api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
alembic = "^1.12.0"
pydantic = "^2.4.2"
sqlalchemy = "^2.0.21"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
fastapi-csrf-protect = "^0.3.2"
pyjwt = "^2.8.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
